mc_pattern.hpp created
